requirements.txt
discord.py>=2.3,<3
aiohttp>=3.9,<4
apscheduler>=3.10,<4
